create placeholder contacts frontend page and fix login flow. Remove old dashboard initial page
