+++ app/Http/Controllers/ContactController.php
@@ -3,8 +3,14 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Inertia\Inertia;
 
 class ContactController extends Controller
 {
-    //
+
+
+    public function show()
+    {
+        return Inertia::render('contacts/index');
+    }
 }
